Refactor BankManagement component and DragFileComponent
This commit refactors the BankManagement component by adding a useState hook for the 'file' state and an useEffect hook to reset the 'file' state when the 'Banco' parameter changes. It also updates the DragFileComponent by extracting the file handling logic into a separate function and using it in both the handleDrop and handleFileChange functions. Additionally, it adds a position-absolute class to the fileInput-container div and a position-absolute class to the ToastMessage component. Finally, it updates the SideBarMenuComponent by setting the maxHeight of the submenu based on the number of links.